+++ test-app/src/locales.js
@@ -378,6 +378,7 @@ const es_translation = {
     "force-atlas2": "force-atlas2",
     "radial": "radial",
     "treedb-tree": "árbol",
+    "compact-tree": "árbol compacto",
     "node labels": "Etiquetas de los nodos",
     "nodes": "Nodos",
     "full": "Completo",